requirements.txt
numpy==1.24.1
opencv_python==4.7.0.68
google-cloud-vision==3.4.1
pygbif==0.6.2
pylibdmtx==0.1.10
python-decouple==3.8
setuptools==58.0.4
mysql-connector-python==8.0.33